Moved TimeseriesTestdata into test utilities. Removed commented out code.
