+++ src/volue/mesh/examples/write_timeseries_points.py
@@ -1,6 +1,7 @@
 import pyarrow as pa
 from datetime import datetime
-from volue.mesh import Connection, Timeseries, eagle_wind
+from volue.mesh import Connection, Timeseries
+from volue.mesh.tests.test_utilities.utilities import eagle_wind
 from volue.mesh.examples import _get_connection_info
